@@ -13,14 +13,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^5.0.3",
-    "@angular/compiler": "^5.0.3",
-    "@angular/core": "^5.0.3",
-    "@angular/forms": "^5.0.3",
-    "@angular/http": "^5.0.3",
-    "@angular/platform-browser": "^5.0.3",
-    "@angular/platform-browser-dynamic": "^5.0.3",
-    "@angular/router": "^5.0.3",
+    "@angular/common": "^6.0.7",
+    "@angular/compiler": "^6.0.7",
+    "@angular/core": "^6.0.7",
+    "@angular/forms": "^6.0.7",
+    "@angular/http": "^6.0.7",
+    "@angular/platform-browser": "^6.0.7",
+    "@angular/platform-browser-dynamic": "^6.0.7",
+    "@angular/router": "^6.0.7",
     "@angular2-material/button": "^2.0.0-alpha.8-3",
     "@angular2-material/button-toggle": "^2.0.0-alpha.8-3",
     "@angular2-material/card": "^2.0.0-alpha.8-3",
@@ -40,41 +40,43 @@
     "@angular2-material/tabs": "^2.0.0-alpha.8-3",
     "@angular2-material/toolbar": "^2.0.0-alpha.8-3",
     "@angular2-material/tooltip": "^2.0.0-alpha.8-3",
-    "angular-in-memory-web-api": "^0.5.1",
-    "angularfire2": "^5.0.0-rc.4",
+    "angular-in-memory-web-api": "^0.6.0",
+    "angularfire2": "^5.0.0-rc.11",
     "bootstrap": "^4.1.1",
-    "core-js": "^2.5.1",
-    "firebase": "^4.6.2",
-    "firebaseui": "^2.5.1",
+    "core-js": "^2.5.7",
+    "firebase": "^5.2.0",
+    "firebaseui": "^3.1.1",
     "font-awesome": "^4.7.0",
-    "rxjs": "^5.5.2",
-    "systemjs": "0.20.19",
+    "rxjs": "^6.2.1",
+    "rxjs-compat": "^6.2.1",
+    "systemjs": "0.21.4",
     "ts-helpers": "^1.1.2",
-    "zone.js": "^0.8.18"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/animations": "^5.0.3",
-    "@angular/cdk": "^5.0.0-rc.2",
+    "@angular-devkit/build-angular": "~0.6.8",
+    "@angular/animations": "^6.0.7",
+    "@angular/cdk": "^6.3.2",
     "@angular/cli": "^6.0.8",
-    "@angular/compiler-cli": "^5.0.3",
-    "@angular/flex-layout": "^2.0.0-beta.10-4905443",
-    "@angular/material": "^5.0.0-rc.2",
+    "@angular/compiler-cli": "^6.0.7",
+    "@angular/flex-layout": "^6.0.0-beta.16",
+    "@angular/material": "^6.3.2",
     "@types/hammerjs": "^2.0.35",
-    "@types/jasmine": "2.8.2",
-    "@types/node": "^8.0.53",
-    "codelyzer": "~4.0.1",
+    "@types/jasmine": "2.8.8",
+    "@types/node": "^10.5.2",
+    "codelyzer": "~4.4.2",
     "hammerjs": "^2.0.8",
-    "jasmine-core": "2.8.0",
+    "jasmine-core": "3.1.0",
     "jasmine-spec-reporter": "4.2.1",
-    "karma": "1.7.1",
+    "karma": "2.0.4",
     "karma-chrome-launcher": "^2.2.0",
     "karma-cli": "^1.0.1",
-    "karma-jasmine": "^1.1.0",
+    "karma-jasmine": "^1.1.2",
     "karma-remap-istanbul": "^0.6.0",
-    "protractor": "~5.2.0",
-    "ts-node": "3.3.0",
-    "tslint": "^5.8.0",
-    "typescript": "~2.6.2",
-    "@angular-devkit/build-angular": "~0.6.8"
+    "protractor": "~5.3.2",
+    "ts-node": "7.0.0",
+    "tslint": "^5.10.0",
+    "typescript": "^2.7.2",
+    "webpack": "^4.15.1"
   }
 }
